Implement event-driven alert system: add documentation, define event types, and enhance alert handling across services.
